1. signal handling
