+++ cmd/container-use/main.go
@@ -4,10 +4,7 @@ import (
 	"context"
 	_ "embed"
 	"fmt"
-	"io"
 	"os"
-	"os/signal"
-	"runtime"
 	"syscall"
 
 	"github.com/charmbracelet/fang"
@@ -26,9 +23,7 @@ Each environment runs in its own container with dedicated git branches.`,
 
 func main() {
 	ctx := context.Background()
-	sigusrCh := make(chan os.Signal, 1)
-	signal.Notify(sigusrCh, syscall.SIGUSR1)
-	go handleSIGUSR(sigusrCh)
+	setupSignalHandling()
 
 	if err := setupLogger(); err != nil {
 		fmt.Fprintf(os.Stderr, "Failed to setup logger: %v\n", err)
@@ -54,20 +49,6 @@ func main() {
 	); err != nil {
 		os.Exit(1)
 	}
-}
-
-func handleSIGUSR(sigusrCh <-chan os.Signal) {
-	for sig := range sigusrCh {
-		if sig == syscall.SIGUSR1 {
-			dumpStacks()
-		}
-	}
-}
-
-func dumpStacks() {
-	buf := make([]byte, 1<<20) // 1MB buffer
-	n := runtime.Stack(buf, true)
-	io.MultiWriter(logWriter, os.Stderr).Write(buf[:n])
 }
 
 func suggestEnvironments(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
