+++ cmd/container-use/signal_windows.go
@@ -1,0 +1,8 @@
+//go:build windows
+
+package main
+
+// On Windows, SIGUSR1 is not available, so we provide a no-op implementation
+func setupSignalHandling() {
+	// No special signal handling on Windows
+}
